Replace the old AptConnection, update device drivers to use the new AptConnection and simplify polling.
