+++ src/pnpq/devices/waveplate_thorlabs_k10cr1.py
@@ -1,3 +1,4 @@
+import atexit
 import threading
 import time
 from abc import ABC, abstractmethod
@@ -8,7 +9,7 @@ from typing import Any, cast
 import structlog
 from pint import Quantity
 
-from ..apt.connection import AptConnection
+from ..apt.connection import AbstractAptConnection
 from ..apt.protocol import (
     Address,
     AptMessage_MGMSG_HW_START_UPDATEMSGS,
@@ -39,6 +40,7 @@ from ..apt.protocol import (
     LimitSwitch,
     StopMode,
 )
+from ..events import Event
 from ..units import pnpq_ureg
 
 
@@ -215,24 +217,28 @@ class AbstractWaveplateThorlabsK10CR1(ABC):
 class WaveplateThorlabsK10CR1(AbstractWaveplateThorlabsK10CR1):
     _chan_ident = ChanIdent.CHANNEL_1
 
-    connection: AptConnection
+    connection: AbstractAptConnection = field()
     home_on_init: bool = field(default=True)
 
-    # Polling threads
-    tx_poller_thread: threading.Thread = field(init=False)
-    tx_poller_thread_lock: threading.Lock = field(default_factory=threading.Lock)
+    # Polling thread
+    _poller_thread: threading.Thread = field(init=False)
+    _stop_poller_event: threading.Event = field(
+        default_factory=threading.Event, init=False
+    )
 
     log = structlog.get_logger()
 
     def __post_init__(self) -> None:
+        atexit.register(self._stop_poller)
+
         # Start polling thread
         object.__setattr__(
             self,
-            "tx_poller_thread",
-            threading.Thread(target=self.tx_poll, daemon=True),
+            "_poller_thread",
+            threading.Thread(target=self._poller),
         )
 
-        self.tx_poller_thread.start()
+        self._poller_thread.start()
 
         # Send autoupdate
         self.connection.send_message_no_reply(
@@ -264,37 +270,6 @@ class WaveplateThorlabsK10CR1(AbstractWaveplateThorlabsK10CR1):
             self.log.info(
                 "[Waveplate] Device is not homed, but skipping homing on setup because home_on_init is set to False.",
             )
-
-    # Polling thread for sending status update requests
-    def tx_poll(self) -> None:
-        with self.tx_poller_thread_lock:
-            while True:
-                self.connection.send_message_unordered(
-                    AptMessage_MGMSG_MOT_ACK_USTATUSUPDATE(
-                        destination=Address.GENERIC_USB,
-                        source=Address.HOST_CONTROLLER,
-                    )
-                )
-                # If we are currently waiting for a reply to a message
-                # we sent, poll every 0.2 seconds to ensure a
-                # relatively quick response to state changes that we
-                # observe using status update messages. If we are not
-                # waiting for a reply, poll at least once every second
-                # to reduce the amount of noise in logs.
-                #
-                # The tx_ordered_sender thread can request a faster
-                # update by setting the
-                # tx_ordered_sender_awaiting_reply event.
-                if self.connection.tx_ordered_sender_awaiting_reply.is_set():
-                    time.sleep(0.2)
-                else:
-                    # The documentation for
-                    # MGMSG_MOT_ACK_USTATUSUPDATE suggests that it
-                    # should be sent at least once a second. This will
-                    # probably send slightly less frequently than once
-                    # a second, so, if we start having issues, we
-                    # should decrease this interval.
-                    self.connection.tx_ordered_sender_awaiting_reply.wait(0.9)
 
     def set_channel_enabled(self, enabled: bool) -> None:
         if enabled:
@@ -591,3 +566,37 @@ class WaveplateThorlabsK10CR1(AbstractWaveplateThorlabsK10CR1):
         )
         assert isinstance(status_message, AptMessage_MGMSG_MOT_GET_STATUSUPDATE)
         return status_message.status.HOMED
+
+    def _poller(self) -> None:
+        """Intended to be run in a background thread.
+
+        Constantly send ``ACK_USTATUSUPDATE`` messages to keep the
+        connection alive; some parts of the official manual suggest
+        that if this ``ACK`` message is not sent at regular intervals,
+        the device will assume the connection is closed and stop
+        responding to commands.
+
+        Generally speaking, this thread will always exit with an
+        exception of some kind; the exception is only logged at debug
+        level to avoid cluttering the log, but if status update
+        messages are mysteriously failing to appear, this is a good
+        place to start.
+
+        """
+        try:
+            while not self._stop_poller_event.is_set():
+                time.sleep(0.9)
+                self.connection.send_message_unordered(
+                    AptMessage_MGMSG_MOT_ACK_USTATUSUPDATE(
+                        destination=Address.GENERIC_USB,
+                        source=Address.HOST_CONTROLLER,
+                    )
+                )
+        except BaseException as e:  # pylint: disable=W0718
+            self.log.debug(event=Event.APT_POLLER_EXIT, exc_info=e)
+
+    def _stop_poller(self) -> None:
+        """Called by atexit when the Python interpreter is exiting to
+        stop the poller thread."""
+        self._stop_poller_event.set()
+        self._poller_thread.join()
